app: more responsive lock popover.
First of all, let's use gtk_widget_show|hide() instead of
gtk_popover_popup|down() because these GtkPopover API sometimes "block"
when clicking very fast on the item list, sometimes up to 2 or 3 seconds
(at least it feels like seconds, I haven't actually timed it!). I'm sure
the ones who developed it thought it was nice to have slow popping
dialogs, but when working fast, this can only be frustrating).

The second responsiveness change is that when clicking out of useful
popover area, it should not only close, but the event should be passed
onto the item tree view when relevant. This allows for instance to
directly click on an eye cell to change a layer visibility without
having to click twice.